Move pi-topOS version checking into its own subdirectory
