+++ pt_os_web_portal/os_updater/manager.py
@@ -1,8 +1,6 @@
 from datetime import date, datetime
 
 from pitop.common.logger import PTLogger
-from pitop.common.pt_os import get_pitopOS_info
-from requests import get
 
 from ..backend.helpers.modules import get_apt
 from ..config_manager import ConfigManager
@@ -49,44 +47,6 @@ class InstallProgress(apt.progress.base.InstallProgress):  # type: ignore
 
     def update_interface(self):
         apt.progress.base.InstallProgress.update_interface(self)
-
-
-# TODO: Move/rename this for 'pi-topOS version check'
-def check_relevant_os_updates():
-    URL = "https://backend-test.pi-top.com/utils/v1/OS/checkUpdate"
-    BUILD_INFO_TO_API_LOOKUP = {
-        "build_os_version": "currentOSVersion",
-        "build_commit": "buildCommit",
-    }
-
-    def build_info_query_params():
-        build_info = get_pitopOS_info()
-        build_info_dict = build_info.__dict__ if build_info else {}
-        return {
-            BUILD_INFO_TO_API_LOOKUP.get(key): value
-            for key, value in build_info_dict.items()
-            if value and BUILD_INFO_TO_API_LOOKUP.get(key)
-        }
-
-    url_query_dict = build_info_query_params()
-    PTLogger.info(
-        f"Checking if there are major OS updates - sending request to {URL} with {url_query_dict}"
-    )
-    data = {
-        "shouldBurn": False,
-        "requireBurn": False,
-        "latestOSVersion": "",
-        "update": False,
-    }
-    try:
-        response = get(URL, url_query_dict, timeout=5).json()
-        PTLogger.info(f"Response was: {response}")
-        for k, v in response.items():
-            data.update({k: v})
-    except Exception as e:
-        PTLogger.warning(f"{e}")
-    finally:
-        return data
 
 
 class OSUpdateManager:
